alertlistener/requirements.txt
# Function dependencies, for example:
google-cloud-datastore==1.8.0
